### 2022-10-14 #4853 @kthyng
Hi @grburgess and thanks for your submission. I think the problem compiling your paper is that you missing your affiliation. See details [here](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).

Also I don't see instructions for installation — are they missing?

Finally, all of our relevant editors are fully occupied at the moment so I'll add this submission to the waitlist. Thanks for your patience.
