### 2022-10-14 #4853 @grburgess
The pdf seemed to work with the GitHub action
