### 2022-10-27 #4853 @dfm
@grburgess — I wanted to provide a quick update here since I've now taken over as managing editor of this JOSS "track". Thanks again for your submission! All of the suitable editors are still working at capacity, but I'm working to get someone assigned ASAP and I'll keep you posted. Thanks for your patience!
