### 2022-11-25 #4969 @editorialbot
[REVIEW]: ronswanson: Building Table Models for 3ML
**Submitting author:** @grburgess (<a href="http://orcid.org/0000-0003-3345-9515">J. Michael Burgess</a>)
**Repository:** https://github.com/grburgess/ronswanson
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.10
