### 2022-12-05 9a8c49c
add ability to construct key files for only missing data
