### 2022-12-05 0969b6b
faster solution
