### 2022-12-15 0de58ec
turn glob into list
