### 2022-12-20 6e694f8
add ability to replace non-finite values
