### 2022-12-20 2a282f2
added warning when reading non-finit values
