### 2023-01-02 8bfe192
added some quality checks
