+++ ronswanson/utils/__init__.py
@@ -1,2 +1,3 @@
 from .configuration import ronswanson_config, show_configuration
 from .color import Colors
+from .quailty_checks import check_for_non_contiguous_spectra
