### 2023-01-02 #4969 @dfm
@volodymyrss — Great question! As I [mentioned briefly above](https://github.com/openjournals/joss-reviews/issues/4969#issuecomment-1327682545), the usual workflow is to open issues on the [project repo](https://github.com/grburgess/ronswanson) (it is useful to include a link back to this thread) with any suggested changes. You're also welcome to list smaller things directly here in this thread, but that can get things a little cluttered. Let me know if you have further questions!
