### 2023-01-03 grburgess/ronswanson#2 @volodymyrss
No such file or directory: '/opt/conda/lib/python3.9/site-packages/ronswanson/data/log_theme.ini'
* ronswanson version: 0.2.5  (pip install ronswanson as of right now)
* Python version: 3.9
* Operating System: bullseye/sid

### Description

Tried to follow the example in the doc http://jmichaelburgess.com/ronswanson/notebooks/ronswanson.html#Example-with-a-Band-function

in relation to https://github.com/openjournals/joss-reviews/issues/4969

### What I Did

```python
FileNotFoundError                         Traceback (most recent call last)
Input In [3], in <cell line: 5>()
      3 import numpy as np
      4 from astromodels import Band
----> 5 import ronswanson as dukesilver
      8 class BandSimulation(dukesilver.Simulation):
      9     def __init__(
     10         self,
     11         simulation_id: int,
   (...)
     14         out_file: str,
     15     ) -> None:

File /opt/conda/lib/python3.9/site-packages/ronswanson/__init__.py:9, in <module>
      6 __email__ = 'jburgess@mpe.mpg.de'
      8 from ._version import get_versions
----> 9 from .database import Database
     10 from .simulation import Simulation
     11 from .simulation_builder import SimulationBuilder

File /opt/conda/lib/python3.9/site-packages/ronswanson/database.py:12, in <module>
      9 from astromodels.utils.logging import silence_console_log
     10 from tqdm.auto import tqdm
---> 12 from ronswanson.grids import Parameter, ParameterGrid
     13 from ronswanson.utils.color import Colors
     15 from .utils.logging import setup_logger

File /opt/conda/lib/python3.9/site-packages/ronswanson/grids.py:9, in <module>
      6 import numpy as np
      7 import yaml
----> 9 from .utils.logging import setup_logger
     11 log = setup_logger(__name__)
     14 @dataclass(frozen=True)
     15 class EnergyGrid:

File /opt/conda/lib/python3.9/site-packages/ronswanson/utils/logging.py:25, in <module>
     16         return logRecord.levelno != self.__level
     19 _console_formatter = logging.Formatter(
     20     ' %(message)s',
     21     datefmt="%H:%M:%S",
     22 )
---> 25 mytheme = Theme().read(get_path_of_data_file("log_theme.ini"))
     26 console = Console(theme=mytheme)
     28 ronswanson_console_log_handler = RichHandler(
     29     level=ronswanson_config.logging.level,
     30     rich_tracebacks=True,
     31     markup=True,
     32     console=console,
     33 )

File /opt/conda/lib/python3.9/site-packages/rich/theme.py:69, in Theme.read(cls, path, inherit)
     58 @classmethod
     59 def read(cls, path: str, inherit: bool = True) -> "Theme":
     60     """Read a theme from a path.
     61
     62     Args:
   (...)
     67         Theme: A new theme instance.
     68     """
---> 69     with open(path, "rt") as config_file:
     70         return cls.from_file(config_file, source=path, inherit=inherit)

FileNotFoundError: [Errno 2] No such file or directory: '/opt/conda/lib/python3.9/site-packages/ronswanson/data/log_theme.ini'
```

https://renkulab.io/projects/vladimir.savchenko/ronswanson-review/files/blob/Untitled.ipynb
