+++ setup.cfg
@@ -29,6 +29,7 @@ project_urls =
 packages = find:
 install_requires =
     numpy
+    matplotlib
     pyyaml
     h5py
     joblib
