+++ _sources/notebooks/api/ronswanson.utils.rst.txt
@@ -76,6 +76,14 @@ ronswanson.utils.package\_data module
    :undoc-members:
    :show-inheritance:
 
+ronswanson.utils.quailty\_checks module
+---------------------------------------
+
+.. automodule:: ronswanson.utils.quailty_checks
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 Module contents
 ---------------
 
