### 2023-01-03 grburgess/ronswanson#2 @volodymyrss
Thanks it works now. I wonder how it works in relation to https://github.com/openjournals/joss-reviews/issues/4969 , it has to be somehow well-defined version which is being reviewed. As it's iterative, it would be the master branch. But some commands in the doc would not work until the version is uploaded. Maybe this is not imporatant. It's more of a question to @dfm.
