### 2023-01-03 grburgess/ronswanson#2 @volodymyrss
> @volodymyrss the best approach is to review the development version of the software and then we always tag a new release at the end of the review to archive the reviewed version.

Ok, I see, the comments/issues can be on master.

But I suppose the approval should be on a fixed revision, right? Else there can be a change on the master branch which is not approved.
How do I track what I am approving?

In principle, this is the case not only for final approval but for every qualification.

E.g. if I check "[ ] Installation: Does installation proceed as outlined in the documentation?", how do I indicate that it this particular version which proceeds, but not the other one?
