### 2023-01-03 grburgess/ronswanson#3 @volodymyrss
Do the authors describe how this software compares to other commonly-used packages?
But do the authors really describe how this software compares to other commonly-used packages?

https://github.com/openjournals/joss-reviews/issues/4969
