### 2023-01-03 grburgess/ronswanson#3 @grburgess
I wasn't really aware if there was something similar.
