### 2023-01-03 grburgess/ronswanson#3 @volodymyrss
According to the [guide](https://joss.readthedocs.io/en/latest/review_checklist.html):

>  A JOSS review is generally considered incomplete until the reviewer has checked off all of their checkboxes.

and there is this one:

> [ ] State of the field: Do the authors describe how this software compares to other commonly-used packages?

So technically I can not "generally" proceed until this is addressed.

On other other hand, in principle, useful software does not need to have such a section to remain useful. But in an ordinary paper one could require it.

What do we do here, @dfm?

Do we expect something like "to the best of  the author's knowledge this work is completely unprecedented"?
