### 2023-01-03 grburgess/ronswanson#3 @dfm
I generally expect that a JOSS paper should include some discussion of the software ecosystem context that the project exists within. If there truly aren't any similar projects, then it's fine to check that box, but I would generally expect there to be some comparable software that should be discussed for context. These comparisons certainly don't need to be in-depth, but should be included if at all possible!
