### 2023-01-03 grburgess/ronswanson#3 @grburgess
I could definitely add something about other spectral fitting software existing (xspec, isis, etc.) but wouldn't this be more related to the fitting package than the table building procedure? I'm unaware of anything other than one off solutions for this problem which is why the package was designed. But if there are others out there I'm happy to cite them.
