### 2023-01-03 grburgess/ronswanson#3 @volodymyrss
Ok then, we need a very brief insight from @grburgess on what are other at least remotely comparable solutions.
Or a statement that there are none.
I suppose in a narrow sense, just for `astromodels` - there are probably none.
But since people achieve something similar in other situations, it could be useful.
