### 2023-01-03 grburgess/ronswanson#3 @grburgess
@volodymyrss I do see this: https://github.com/mbursa/xspec-table-models which is a wrapper to generate an XSPEC fits file, but from an already existing simulation. Either way, I will write something up in a few moments.
