+++ paper.bib
@@ -47,4 +47,62 @@ volume = {1507},
 local-url = {file://localhost/Users/jburgess/Documents/Papers%20Library/2015/Vianello/Vianello_2015_The%20Multi-Mission%20Maximum%20Likelihood%20framework%20(3ML).pdf}
 }
 
-@techreport{mpi, author = {Forum, Message P}, title = {MPI: A Message-Passing Interface Standard}, year = {1994}, publisher = {University of Tennessee}, address = {USA}, abstract = {The Message Passing Interface Forum (MPIF), with participation from over 40 organizations, has been meeting since November 1992 to discuss and define a set of library standards for message passing. MPIF is not sanctioned or supported by any official standards organization. The goal of the Message Passing Interface, simply stated, is to develop a widely used standard for writing message-passing programs. As such the interface should establish a practical, portable, efficient and flexible standard for message passing. , This is the final report, Version 1.0, of the Message Passing Interface Forum. This document contains all the technical features proposed for the interface. This copy of the draft was processed by LATEX on April 21, 1994. , Please send comments on MPI to mpi-comments@cs.utk.edu. Your comment will be forwarded to MPIF committee members who will attempt to respond.} }
+@techreport{mpi, author = {Forum, Message P}, title = {MPI: A Message-Passing
+                  Interface Standard}, year = {1994}, publisher = {University of
+                  Tennessee}, address = {USA}, abstract = {The Message Passing
+                  Interface Forum (MPIF), with participation from over 40
+                  organizations, has been meeting since November 1992 to discuss
+                  and define a set of library standards for message
+                  passing. MPIF is not sanctioned or supported by any official
+                  standards organization. The goal of the Message Passing
+                  Interface, simply stated, is to develop a widely used standard
+                  for writing message-passing programs. As such the interface
+                  should establish a practical, portable, efficient and flexible
+                  standard for message passing. , This is the final report,
+                  Version 1.0, of the Message Passing Interface Forum. This
+                  document contains all the technical features proposed for the
+                  interface. This copy of the draft was processed by LATEX on
+                  April 21, 1994. , Please send comments on MPI to
+                  mpi-comments@cs.utk.edu. Your comment will be forwarded to
+                  MPIF committee members who will attempt to respond.} }
+
+@INPROCEEDINGS{gammapy,
+author = {{Deil}, C. and {Zanin}, R. and {Lefaucheur}, J. and {Boisson}, C. and
+         {Khelifi}, B. and {Terrier}, R. and {Wood}, M. and {Mohrmann}, L. and
+         {Chakraborty}, N. and {Watson}, J. and {Lopez-Coto}, R. and
+         {Klepser}, S. and {Cerruti}, M. and {Lenain}, J.~P. and {Acero}, F. and
+         {Djannati-Ata{\"\i}}, A. and {Pita}, S. and {Bosnjak}, Z. and
+         {Trichard}, C. and {Vuillaume}, T. and {Donath}, A. and
+         {Consortium}, CTA and {King}, J. and {Jouvin}, L. and {Owen}, E. and
+         {Sipocz}, B. and {Lennarz}, D. and {Voruganti}, A. and
+         {Spir-Jacob}, M. and {Ruiz}, J. Enrique and {Arribas}, M. Paz},
+title = "{Gammapy - A prototype for the CTA science tools}",
+keywords = {Astrophysics - Instrumentation and Methods for Astrophysics},
+booktitle = {35th International Cosmic Ray Conference (ICRC2017)},
+year = 2017,
[…]
