### 2023-01-03 8d8fe5b
remove double citation
