### 2023-01-03 grburgess/ronswanson#4 @volodymyrss
Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
Are the community guidelines?

https://github.com/openjournals/joss-reviews/issues/4969
