### 2023-01-03 1250b7a
fix console scripts
