### 2023-01-03 grburgess/ronswanson#2 @volodymyrss
> @volodymyrss Good question! We typically are not strict about that for JOSS reviews - you're definitely not expected to go through the full checklist a second time when there is a final version minted. I'd say that this is partly because of logistics (the scope is too broad, and it's hard to synchronize the two reviews, for example), but also partly because the point of a JOSS review is more to provide constructive feedback than to bring a specific judgement. From my perspective, it's totally sufficient to make sure that at the time you're going through it, the development branch satisfies the criteria in the checklist.

I see, thanks.

> That being said, if there are elements that are brittle or give you other reasons to be concerned, it's good to note that, and when I'm reviewing I do sometimes quickly go through the checklist one last time before signing off, but that's certainly not required! Please let me know if you have further questions - I'm happy to continue this conversation!

Installation is quite likely to break in the future, especially if the requirements are not frozen. I suppose this is normal and not a fault of the software.
Evolving requirements may even change the behavior, or at least performance.
If we are concerned here with making the review reproducible and verifiable, I'd attach some form of environment snapshot, with all fixed versions, in a container.

But at minimum, I could point out the revision I am looking at when making a statement.
Maybe your bot could make use of this information in the future.
