### 2023-01-04 grburgess/ronswanson#2 @dfm
Great - totally agreed! And thanks for the suggestion - I'll bring this point up with the other JOSS editors to see how we'd like to handle these versioning issues that you've brought up. Thanks again!!
