### 2023-01-08 grburgess/ronswanson#6 @cosimoNigro
Broadening the scope of ronswanson: building table models for other software?
This issue is related [to ronswanson's review for the journal of the open source software](https://github.com/openjournals/joss-reviews/issues/4969), specifically to the item:
> Substantial scholarly effort: Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)

The software is quite young, [it started in summer 2022](https://github.com/grburgess/ronswanson/graphs/contributors). I do not doubt the amount of work it went in, nor that it will be cited in future; but the fact that it produces template models only for one specific software (`3ML`), I think makes its scope a bit narrow. Would it be too much work to allow to generate also template models for other packages that have serialisable models (also via `.yaml` files), e.g. `gammapy`? I think it would be a huge advantage for this work and it would significantly broaden its scope and usage.

On the other hand, maybe the authors consider this outside of their original intention. It would be good to hear their opinion on this. From a quick look at the documentation, it seems to me that the database with the model and its parameters can be dumped in a `3ML` model. Can't a functionality be added to do the same with `gammapy`? Are there other fundamental routines in `ronswanson` using `3ML`?

If finally generating `gammapy`-compatible models is judged too much, then maybe a paragraph in the paper can outline future possible expansions of `ronswanson` to other modelling packages with serialisable models?
