### 2023-01-08 grburgess/ronswanson#7 @cosimoNigro
Docs: separate the intro and make a clearer case for ronswanson
This issue is related [to ronswanson's review for the journal of the open source software](https://github.com/openjournals/joss-reviews/issues/4969), specifically to the item:

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

Now in the docs there is a single `Intro` page with a brief statement of need, installation instructions and a first example.
On the other hand, in the main doc page there is a brief introductory paragraph just stating:

> This tool provides an easy interface to construct 3ML table models from computational expensive simulations on HPC systems. Why? Because it is your right to do so!

Though a conversational approach is always welcome in documentations, maybe the authors can better detail here the problem the software aims to solve, add some physical applications (what is the case of a simulation / analysis for which `ronswanson` would be useful?) and also for whom this software is helpful?

Maybe it would be better to have the installation instructions separated from the first example.
They can go in the main page as well.
