### 2023-01-09 grburgess/ronswanson#8 @cosimoNigro
Improving the example in the documentation - eventually adding more
This issue is related [to ronswanson's review for the journal of the open source software](https://github.com/openjournals/joss-reviews/issues/4969), specifically to the item:

> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

I tried to re-run the `Intro` example myself, but - to be honest - it took me quite some time to understand what the different steps of the tutorial represent. I feel that the explanation can be improved.

Moreover, reading the paper draft, I got the idea that the unique feature provided by this package was that:
> one can create a so-called template or table model by evaluating the simulation on a grid of its parameters, and use interpolation on the output which allows for the simulated model to compared with data via the likelihood in a reasonable amount of time.

But in this tutorial (the only one in the documentation) this is not fully shown.
We are only shown a simple analytical function that is evaluated on a grid of parameters and whose values vs energy are stored in a database. I understand that the band function is used for simplicity (and that `ronswanson` might be especially useful for more sophisticated models) but we are not shown how this model database can speed up a statistical analysis leveraging the interpolation of the model outputs that the author discusses. And maybe this is non trivial to show, but the current example does not show the particular feature advertised in the paper.

Maybe it would be good to add a statistical evaluation relying on the models in the database in this tutorial, to show the advantages of using this framework? Or maybe - even better - one could add another tutorial notebook illustrating the  speed-up in the statistical calculation with the database models?

Is it possible to create only energy-dependent models - f(E), can other type of models be created? This is not clear.

P.S. I would not use the term "simulation" for a simple analytical function evaluated on a grid of parameters.
