### 2023-01-09 #4969 @cosimoNigro
Hello @grburgess, @dfm, @volodymyrss,

Apologies for the delay in the review, I am really sorry for it.

**Find here a resume of the issues I opened in the software repository and the corresponding items in the Review checklist left unchecked.**

### General checks
> Substantial scholarly effort: Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)

https://github.com/grburgess/ronswanson/issues/6

### Documentation
> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

https://github.com/grburgess/ronswanson/issues/7

> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

https://github.com/grburgess/ronswanson/issues/8

### Software paper
- Line 5: "`ronswanson` provides a simple to use framework" ... -> `ronswanson` is a python package providing a simple-to-use framework...;

- Line 7:  "This allows for the construction of fast evaluating interpolation table of expensive computer simulations of physical models which can then be fitted to data in a reasonable amount of time." -> confused, would be made more clear by moving here the sentence that goes from line 16 to line 20. That part of the text clearly states the problem at hand;

- Line 12: "Users can easily prototype their pipeline on multi-core workstations and then switch to a multi-node HPC system" -> so `ronswanson` can also generate scripts to parallelise the statistical analysis, beside the model evaluation? This is not clear;

- Line 16: "Spatio-spectral fitting" -> Spectro-morphological?

- Line 28: "While astromodels provides user friendly factories" -> user-friendly factories;

- Line 29: "the workflow for using these factories on desktop workstations or HPC systems can be complex." -> yet not clear if `ronswanson` can provide also script to parallelise the statistical analysis, beside the table model creation;

 - Line 35, 36: "it's output", "it's virtual run" -> its;

- Line 36-44: it would be nice if readers could visualise the process of model creation from the parameters. Perhaps adding a simple plot / scheme - [the grid with the parameters values displayed in the documentation](http://jmichaelburgess.com/ronswanson/notebooks/ronswanson.html#The-Database) might be a good start;

- Line 45: "Additionally, the energy grid corresponding to the evaluation of each of the simulation outputs must be specified in this file." ->  why is the energy specified, can `ronswanson` only create tables for spectral models?

- Line 52: "This automatically generates all the required python and SLURM scripts required for the construction of the table model" -> do you mean that they actually allow to evaluate the model for the different parameters and collect the different outputs. When does the interpolation that is supposed to speed up the stat. analysis happen? At this moment? After the table creation.

- Line 60: "an table model" -> a table model;

- Line 63: "that certain fits" -> that certain statistical analyses?

- Line 67: full stop missing.

P.S. In all citations the parentheses immediately follow the text without a space, is this a problem with `\citep`?
