### 2023-01-09 grburgess/ronswanson#6 @grburgess
@cosimoNigro I think this is a fair point and could be added on at some point, but this effort may be irrelevant as there is already an existing bridge in 3ML to use gammapy. Specifically, there is a plugin which pumps the models in astromodels into the gammapy likelihood allowing all the fitting capabilities (and multi-instrumental bridges) in 3ML to intefface directly with instruments that utilize gammapy as their likelihood interface:

This work is still preliminary and there have been some small discussions on perhaps having a single modeling language like astromodels in future versions of gammapy to ease the effort currently underway in developing the gammapy plugin. So for now I would like to push this off to a future version... if it will even be needed at all.

On a side note, is there a similar model template model factory in gammapy like what is provided in astromodels? I could not find this in the docs, but it would make having a future interface easier.

Does this satisfy your concern?
