### 2023-01-09 grburgess/ronswanson#7 @grburgess
@cosimoNigro  This is a good point. I will clean up the docs and describe the problem more technically. I think I will also link in the docs from astromodels for further reading on the technical details of the template models as `ronswanson` is more focused on the the construction of the base data.
