### 2023-01-09 grburgess/ronswanson#8 @grburgess
@cosimoNigro I think this could be done.. there are working examples for real simulations (the motivation behind the work) that are unfortunately closed source at the moment as they are from proprietary blazar spectral code that I do not have permission to open-source at the moment.

Would you prefer something like a mock "simulation" code that is interfaced in? Even in the case of the blazer "simulation" code, the actual calls to the evaluation are python hooks... so there is not a giant leap in going from something like Band function or a power law.
