### 2023-01-09 grburgess/ronswanson#8 @grburgess
On the issue of showing the template model in a fit, would it just be sufficient to link the associated software documentation showing how these tools can be combined?
