### 2023-01-31 grburgess/ronswanson#6 @cosimoNigro
> This work is still preliminary and there have been some small discussions on perhaps having a single modeling language like astromodels in future versions of gammapy to ease the effort currently underway in developing the gammapy plugin. So for now I would like to push this off to a future version... if it will even be needed at all.

Maybe you can find a way to mention this briefly in the paper? Was this "pumping of the astromodels directly in the Gammapy likelihood" presented in some conference already?

> On a side note, is there a similar model template model factory in gammapy like what is provided in astromodels? I could not find this in the docs, but it would make having a future interface easier.

Not that I am aware of.

> Does this satisfy your concern?

As written above, if you can mention this bridge with Gammapy it would be great.
I understand what I was asking is an effort larger than the software scope at this particular moment in time.
