### 2023-01-31 grburgess/ronswanson#8 @cosimoNigro
I does not matter that the model is not an actual physical model but an analytical approximation.

What I would like to see is a statistical application of the model generated and, if possible, the eventual speed-up introduced by having and the model already evaluated on a grid of data.

> On the issue of showing the template model in a fit, would it just be sufficient to link the associated software documentation showing how these tools can be combined?

To which documentation and tools you are referring to?
