### 2023-03-10 c087cb9
make parameter dicts
