### 2023-03-10 e4fed73
fix n_points for LHS
