### 2023-03-20 beb2ed8
switch to using scipy over smt
