### 2023-03-22 grburgess/ronswanson#6 @grburgess
@cosimoNigro I have implemented these into the paper now.
