### 2023-03-22 grburgess/ronswanson#8 @grburgess
@cosimoNigro I have tried to make the connection between the example and a real simulation clearer. As far as fitting, I could put some mock data, but it would make the repo heavier. I have instead linked to the 3ML docs which already cover both fitting and fitting table models. Would this be ok?
