+++ paper.bib
@@ -66,6 +66,23 @@ local-url = {file://localhost/Users/jburgess/Documents/Papers%20Library/2015/Via
                   mpi-comments@cs.utk.edu. Your comment will be forwarded to
                   MPIF committee members who will attempt to respond.} }
 
+
+
+@INPROCEEDINGS{xspec,
+       author = {{Arnaud}, K.~A.},
+        title = "{XSPEC: The First Ten Years}",
+    booktitle = {Astronomical Data Analysis Software and Systems V},
+         year = 1996,
+       editor = {{Jacoby}, George H. and {Barnes}, Jeannette},
+       series = {Astronomical Society of the Pacific Conference Series},
+       volume = {101},
+        month = jan,
+        pages = {17},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1996ASPC..101...17A},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
 @INPROCEEDINGS{gammapy,
 author = {{Deil}, C. and {Zanin}, R. and {Lefaucheur}, J. and {Boisson}, C. and
          {Khelifi}, B. and {Terrier}, R. and {Wood}, M. and {Mohrmann}, L. and
@@ -90,19 +107,4 @@ eprint = {1709.01751},
 primaryClass = {astro-ph.IM},
 adsurl = {https://ui.adsabs.harvard.edu/abs/2017ICRC...35..766D},
 adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-
-@INPROCEEDINGS{xspec,
-       author = {{Arnaud}, K.~A.},
-        title = "{XSPEC: The First Ten Years}",
-    booktitle = {Astronomical Data Analysis Software and Systems V},
-         year = 1996,
-       editor = {{Jacoby}, George H. and {Barnes}, Jeannette},
-       series = {Astronomical Society of the Pacific Conference Series},
-       volume = {101},
-        month = jan,
-        pages = {17},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/1996ASPC..101...17A},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
