+++ paper.bib
@@ -108,3 +108,60 @@ primaryClass = {astro-ph.IM},
 adsurl = {https://ui.adsabs.harvard.edu/abs/2017ICRC...35..766D},
 adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+
+@software{acero_fabio_2023_7734804,
+  author       = {Acero, Fabio and
+                  Aguasca-Cabot, Arnau and
+                  Buchner, Johannes and
+                  Carreto Fidalgo, David and
+                  Chen, Andrew and
+                  Chromey, Alisha and
+                  Contreras Gonzalez, José Luis and
+                  de Bony de Lavergne, Mathieu and
+                  de Miranda Cardoso, José Vinícius and
+                  Deil, Christoph and
+                  Donath, Axel and
+                  Giunti, Luca and
+                  Hinton, James and
+                  Jouvin, Léa and
+                  Khélifi, Bruno and
+                  King, Johannes and
+                  Lefaucheur, Julien and
+                  Lenain, Jean-Philippe and
+                  Linhoff, Maximilian and
+                  López-Coto, Rubén and
+                  Mohrmann, Lars and
+                  Morcuende, Daniel and
+                  Nakashima, Kaori and
+                  Nigro, Cosimo and
+                  Olivera-Nieto, Laura and
+                  Owen, Ellis and
+                  Panny, Sebastian and
+                  Papadopoulos Orfanos, Dimitri and
+                  Paz Arribas, Manuel and
+                  Pintore, Fabio and
+                  Poon, Helen and
+                  Remy, Quentin and
+                  Ruiz, José Enrique and
+                  Siejkowski, Hubert and
+                  Sinha, Atreyee and
+                  Sipőcz, Brigitta M and
+                  Spir-Jacob, Marion and
+                  Terrier, Régis and
+                  Tibaldo, Luigi and
+                  Unbehaun, Tim and
+                  van Eldik, Christopher and
+                  Vuillaume, Thomas and
+                  Weinstein, Amanda and
+                  Wood, Matthew},
+  title        = {Gammapy: Python toolbox for gamma-ray astronomy},
+  month        = mar,
+  year         = 2023,
+  note         = {{If you use this software, please cite it using the 
+                   metadata from this file.}},
+  publisher    = {Zenodo},
+  version      = {v1.0.1},
+  doi          = {10.5281/zenodo.7734804},
+  url          = {https://doi.org/10.5281/zenodo.7734804}
+}
