+++ paper.bib
@@ -110,7 +110,7 @@ adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
 
-@software{acero_fabio_2023_7734804,
+@software{acero,
   author       = {Acero, Fabio and
                   Aguasca-Cabot, Arnau and
                   Buchner, Johannes and
