+++ _sources/notebooks/ronswanson.ipynb.txt
@@ -2,7 +2,7 @@
  "cells": [
   {
    "cell_type": "markdown",
-   "id": "d8225562",
+   "id": "c57cf9d7",
    "metadata": {},
    "source": [
     "# Introduction and demonstration\n",
@@ -18,7 +18,7 @@
   },
   {
    "cell_type": "markdown",
-   "id": "00b2bb0d",
+   "id": "af2a8e77",
    "metadata": {},
    "source": [
     "## Example\n",
@@ -72,7 +72,7 @@
   },
   {
    "cell_type": "markdown",
-   "id": "35ed83a1",
+   "id": "47299096",
    "metadata": {},
    "source": [
     "### The Simulation class\n",
@@ -90,25 +90,25 @@
   {
    "cell_type": "code",
    "execution_count": 1,
-   "id": "85b165c3",
+   "id": "34d659f3",
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2023-03-22T12:44:00.561459Z",
-     "iopub.status.busy": "2023-03-22T12:44:00.561077Z",
-     "iopub.status.idle": "2023-03-22T12:44:02.698306Z",
-     "shell.execute_reply": "2023-03-22T12:44:02.697325Z"
+     "iopub.execute_input": "2023-03-28T16:17:21.896188Z",
+     "iopub.status.busy": "2023-03-28T16:17:21.895420Z",
+     "iopub.status.idle": "2023-03-28T16:17:24.125496Z",
+     "shell.execute_reply": "2023-03-28T16:17:24.124393Z"
     }
    },
    "outputs": [
     {
      "data": {
       "text/html": [
-       "<pre style=\"white-space:pre;overflow-x:auto;line-height:normal;font-family:Menlo,'DejaVu Sans Mono',consolas,'Courier New',monospace\"><span style=\"color: #00ff00; text-decoration-color: #00ff00\">12:44:01 </span><span style=\"color: #af5fd7; text-decoration-color: #af5fd7\">WARNING </span> <span style=\"color: #c6c6c6; text-decoration-color: #c6c6c6; font-weight: bold\"> The naima package is not available. Models that depend on it will not be         </span><a href=\"file:///opt/hostedtoolcache/Python/3.8.16/x64/lib/python3.8/site-packages/astromodels/functions/functions_1D/functions.py\" target=\"_blank\"><span style=\"color: #7f7f7f; text-decoration-color: #7f7f7f\">functions.py</span></a><span style=\"color: #7f7f7f; text-decoration-color: #7f7f7f\">:</span><a href=\"file:///opt/hostedtoolcache/Python/3.8.16/x64/lib/python3.8/site-packages/astromodels/functions/functions_1D/functions.py#48\" target=\"_blank\"><span style=\"color: #7f7f7f; text-decoration-color: #7f7f7f\">48</span></a>\n",
[…]
