### 2023-03-28 #4969 @dfm
@grburgess — Thanks! Can you check [the DOIs listed above](https://github.com/openjournals/joss-reviews/issues/4969#issuecomment-1487316359) and add them to the bibliography if they are correct?
