+++ paper.bib
@@ -16,6 +16,7 @@
 	pages = {44--60},
 	publisher = {Springer Berlin Heidelberg},
 	title = {SLURM: Simple Linux Utility for Resource Management},
+	doi = {10.1007/10968987_3},
 	year = {2003}}
 
 @software{astromodels,
@@ -43,7 +44,8 @@ eprint = {1507.08343},
 url = {http://arxiv.org/abs/1507.08343}, 
 abstract = {{Astrophysical sources are now observed by many different instruments at different wavelengths, from radio to high-energy gamma-rays, with an unprecedented quality. Putting all these data together to form a coherent view, however, is a very difficult task. Each instrument has its own data format, software and analysis procedure, which are difficult to combine. It is for example very challenging to perform a broadband fit of the energy spectrum of the source. The Multi-Mission Maximum Likelihood framework (3ML) aims to solve this issue, providing a common framework which allows for a coherent modeling of sources using all the available data, independent of their origin. At the same time, thanks to its architecture based on plug-ins, 3ML uses the existing official software of each instrument for the corresponding data in a way which is transparent to the user. 3ML is based on the likelihood formalism, in which a model summarizing our knowledge about a particular region of the sky is convolved with the instrument response and compared to the corresponding data. The user can choose between a frequentist analysis, and a Bayesian analysis. In the former, parameters of the model are optimized in order to obtain the best match to the data (i.e., the maximum of the likelihood). In the latter, the priors specified by the user are used to build the posterior distribution, which is then sampled with Markov Chain Monte Carlo or Multinest. Our implementation of this idea is very flexible, allowing the study of point sources as well as extended sources with arbitrary spectra. We will review the problem we aim to solve, the 3ML concepts and its innovative potential.}}, 
 pages = {arXiv:1507.08343}, 
-volume = {1507}, 
+volume = {1507},
+doi = {10.22323/1.312.0130},
 local-url = {file://localhost/Users/jburgess/Documents/Papers%20Library/2015/Vianello/Vianello_2015_The%20Multi-Mission%20Maximum%20Likelihood%20framework%20(3ML).pdf}
 }
 
@@ -106,7 +108,8 @@ archivePrefix = {arXiv},
 eprint = {1709.01751},
 primaryClass = {astro-ph.IM},
 adsurl = {https://ui.adsabs.harvard.edu/abs/2017ICRC...35..766D},
-adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+doi = {10.22323/1.301.0766}
 }
 
 
